Initial
Radiator_rating.m uses data and equations from Compact Heat Exchangers Third Edition, by Kays and London to predict the heat dissipation rate of a radiator, given its dimensions and flow specifications. Currently it has not been verified by experiment, however the numbers it outputs match what I expect it to. 

The dimensions that are currently in the file should be disregarded.